Improve code quality (#20)
* Check go.mod in CI workflow

* Rename test file to match SUT

* Group some packages under internal folder

This improves the clarity of the architecture.

* Add tests for missing data

* Update gitignore
