+++ .gitignore
@@ -1,2 +1,5 @@
 coverage.out
 .env
+
+# Editor files
+.idea
